extract types to prevent circular dependency error
