+++ packages/react-ui/src/app/features/builder/run-list/index.tsx
@@ -10,12 +10,10 @@ import { useQuery } from '@tanstack/react-query';
 import { t } from 'i18next';
 import React from 'react';
 
-import {
-  LeftSideBarType,
-  useBuilderStateContext,
-} from '@/app/features/builder/builder-hooks';
+import { useBuilderStateContext } from '@/app/features/builder/builder-hooks';
 import { flowRunsApi } from '@/app/features/flow-runs/lib/flow-runs-api';
 import { FlowRun, SeekPage } from '@openops/shared';
+import { LeftSideBarType } from '../builder-types';
 
 import { FlowRunCard } from './flow-run-card';
